[ADD/#125] 알람리스트 API domain 모듈 추가
