--- domain/src/main/kotlin/co/orange/domain/entity/response/AlarmItemModel.kt
@@ -1,12 +1,0 @@
-package co.orange.domain.entity.response
-
-data class AlarmItemModel(
-    val alarmId: Long,
-    val alarmCase: String,
-    val title: String,
-    val content: String,
-    val time: String,
-    val isChecked: Boolean,
-    val orderId: String?,
-    val itemId: String?,
-)
