+++ domain/src/main/kotlin/co/orange/domain/entity/response/AlarmListModel.kt
@@ -1,0 +1,16 @@
+package co.orange.domain.entity.response
+
+data class AlarmListModel(
+    val alarmList: List<AlarmItemModel>,
+) {
+    data class AlarmItemModel(
+        val alarmId: Long,
+        val alarmCase: String,
+        val title: String,
+        val content: String,
+        val time: String,
+        val isChecked: Boolean,
+        val orderId: String?,
+        val itemId: String?,
+    )
+}
