+++ domain/src/main/kotlin/co/orange/domain/repository/HomeRepository.kt
@@ -1,5 +1,6 @@
 package co.orange.domain.repository
 
+import co.orange.domain.entity.response.AlarmListModel
 import co.orange.domain.entity.response.HomeModel
 import co.orange.domain.entity.response.SearchResultModel
 
@@ -7,4 +8,8 @@ interface HomeRepository {
     suspend fun getHomeData(): Result<HomeModel>
 
     suspend fun getSearchResult(keyword: String): Result<SearchResultModel>
+
+    suspend fun getAlarmList(): Result<AlarmListModel>
+
+    suspend fun patchToReadAlarm(alarmId: Long): Result<Boolean>
 }
